prepared for 0.0.7 release
